improvements

+++ database/init_db.py
@@ -6,13 +6,15 @@ DB_PATH = BASE_DIR / "data" / "college.db"
 SQL_DIR = BASE_DIR / "sql"
 
 
-def init_db():
+def init_db(drop=True):
     # Ensure data dir exists
     DB_PATH.parent.mkdir(parents=True, exist_ok=True)
 
     # Recreate DB if you want a fresh start each run
-    if DB_PATH.exists():
-        DB_PATH.unlink()
+    if drop:
+        if DB_PATH.exists():
+            print(f"Dropping database at {DB_PATH}")
+            DB_PATH.unlink()
 
     conn = sqlite3.connect(DB_PATH)
     cursor = conn.cursor()
@@ -29,11 +31,28 @@ def init_db():
 
 
 if __name__ == "__main__":
-    import backfill
 
-    print('Initiating Database:')
+    ######
+    YEAR_START = 2000
+    YEAR_END = 2024
+    POLL_IDS = [1, 2]
+    ######
+
+    YEARS = list(range(YEAR_START, YEAR_END + 1))
+
+    import backfill
+    import time
+
+    start_time = time.perf_counter()
+
+    print("Initiating Database:")
     init_db()
-    print('DONE.')
-    print('\nBackfilling DB from ESPN')
-    backfill.main()
-    print('DONE.')
+    print("DONE.")
+
+    print("\nBackfilling DB from ESPN")
+    backfill.main(years=YEARS, poll_ids=POLL_IDS)
+    print("DONE.")
+
+    end_time = time.perf_counter()
+    elapsed = end_time - start_time
+    print(f"\nTotal execution time: {elapsed:.2f} seconds")
